Homework #5 Исправление замечания #2 3. BookAuthorDao и BookGenreDao. DAO - это Data Access Object. Т.е. что-то что работает с объектами. У нас нет объекта BookAuthor или BookGenre, а Book, Author и Genre есть. Соответственно то что сейчас в указанных дао это просто методы BookDao, AuthorDao и GenreDao
