--- homework-05/src/main/java/ru/otus/homework/dao/BookAuthorDaoJdbc.java
@@ -1,97 +1,0 @@
-package ru.otus.homework.dao;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcOperations;
-import org.springframework.stereotype.Repository;
-import ru.otus.homework.domain.Author;
-import ru.otus.homework.domain.Book;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-@Repository
-public class BookAuthorDaoJdbc implements BookAuthorDao {
-
-    private final NamedParameterJdbcOperations jdbc;
-
-    @Autowired
-    public BookAuthorDaoJdbc(NamedParameterJdbcOperations jdbc) {
-        this.jdbc = jdbc;
-    }
-
-    @Override
-    public void insertAuthorsByBookUid(long bookUid, List<Author> authors) {
-        if (authors == null){
-            return;
-        }
-        authors.forEach(author -> {
-            Map<String, Object> params = new HashMap<>(2);
-            params.put("bookUid", bookUid);
-            params.put("authorUid", author.getUid());
-            jdbc.update("insert into tbl_book_author(book_uid, author_uid) values(:bookUid, :authorUid)", params);
-        });
-    }
-
-    @Override
-    public void editAuthorsByBookUid(long bookUid, List<Author> authors) {
-        final Map<String, Object> params = new HashMap<>(1);
-        params.put("bookUid", bookUid);
-        jdbc.update("delete from tbl_book_author where book_uid = :bookUid", params);
-        this.insertAuthorsByBookUid(bookUid, authors);
-    }
-
-    @Override
-    public void deleteAuthorsByBookUid(long bookUid, List<Author> authors) {
-        authors.forEach(author -> {
-            Map<String, Object> params = new HashMap<>(2);
-            params.put("bookUid", bookUid);
-            params.put("authorUid", author.getUid());
-            jdbc.update("delete from tbl_book_author where book_uid = :bookUid and author_uid = :authorUid", params);
-        });
-    }
-
-    @Override
-    public List<Book> getBooksByAuthorUid(long authorUid) {
-        final Map<String, Object> params = new HashMap<>(1);
-        params.put("authorUid", authorUid);
-        return jdbc.query("select b.uid, b.title, b.isbn, b.publication_year from tbl_book_author ba\n" +
-                "join tbl_authors a on ba.author_uid = a.uid\n" +
-                "join tbl_books b on ba.book_uid = b.uid\n" +
-                "where a.uid = :authorUid", params, new BookMapper());
-    }
-
-    @Override
-    public List<Author> getAuthorsByBookUid(long bookUid) {
-        final Map<String, Object> params = new HashMap<>(1);
-        params.put("bookUid", bookUid);
-        return jdbc.query("select a.uid, a.full_name, a.pen_name from tbl_book_author ba\n" +
-                "join tbl_authors a on ba.author_uid = a.uid\n" +
-                "join tbl_books b on ba.book_uid = b.uid\n" +
-                "where b.uid = :bookUid", params, new AuthorMapper());
-    }
-
-    private static class AuthorMapper implements RowMapper<Author> {
-        @Override
-        public Author mapRow(ResultSet resultSet, int i) throws SQLException {
-            long uid = resultSet.getLong("uid");
-            String fullName = resultSet.getString("full_name");
-            String penName = resultSet.getString("pen_name");
-            return new Author(uid, fullName, penName);
-        }
-    }
-
-    private static class BookMapper implements RowMapper<Book> {
-        @Override
-        public Book mapRow(ResultSet resultSet, int i) throws SQLException {
-            long uid = resultSet.getLong("uid");
-            String title = resultSet.getString("title");
-            long isbn = resultSet.getLong("isbn");
-            short publication_year = resultSet.getShort("publication_year");
-            return new Book(uid, title, isbn, publication_year);
-        }
-    }
-}
