+++ homework-05/src/main/java/ru/otus/homework/service/GenreServiceImpl.java
@@ -2,9 +2,7 @@ package ru.otus.homework.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.otus.homework.dao.BookGenreDao;
 import ru.otus.homework.dao.GenreDao;
-import ru.otus.homework.domain.Book;
 import ru.otus.homework.domain.Genre;
 import ru.otus.homework.dto.BookDto;
 import ru.otus.homework.dto.GenreDto;
@@ -15,13 +13,13 @@ import java.util.List;
 public class GenreServiceImpl implements GenreService {
 
     private final GenreDao genreDao;
-    private final BookGenreDao bookGenreDaoDao;
+    private final BookService bookService;
     private final UtilityService utilityService;
 
     @Autowired
-    public GenreServiceImpl(GenreDao genreDao, BookGenreDao bookDao, UtilityService utilityService) {
+    public GenreServiceImpl(GenreDao genreDao, BookService bookService, UtilityService utilityService) {
         this.genreDao = genreDao;
-        this.bookGenreDaoDao = bookDao;
+        this.bookService = bookService;
         this.utilityService = utilityService;
     }
 
@@ -45,8 +43,7 @@ public class GenreServiceImpl implements GenreService {
     @Override
     public GenreDto getByUid(long uid) {
         GenreDto genreDto = new GenreDto(genreDao.getByUid(uid));
-        List<Book> books = bookGenreDaoDao.getBooksByGenreUid(uid);
-        genreDto.setBooks(utilityService.convertToBookDto(books));
+        genreDto.setBooks(bookService.getBooksByGenreUid(uid));
         return genreDto;
     }
 
@@ -54,8 +51,7 @@ public class GenreServiceImpl implements GenreService {
     public List<GenreDto> getAll() {
         List<GenreDto> genreDtos = utilityService.convertToGenreDto(genreDao.getAll());
         genreDtos.forEach(genreDto -> {
-            List<BookDto> bookDtos =
-                    utilityService.convertToBookDto(bookGenreDaoDao.getBooksByGenreUid(genreDto.getUid()));
+            List<BookDto> bookDtos = bookService.getBooksByGenreUid(genreDto.getUid());
             genreDto.setBooks(bookDtos);
 
         });
@@ -65,5 +61,25 @@ public class GenreServiceImpl implements GenreService {
     @Override
     public int count() {
         return genreDao.count();
+    }
+
+    @Override
+    public void insertGenresByBookUid(long bookUid, List<GenreDto> genres) {
+        genreDao.insertGenresByBookUid(bookUid, utilityService.convertToGenreDomain(genres));
+    }
+
+    @Override
+    public void editGenresByBookUid(long bookUid, List<GenreDto> genres) {
+        genreDao.editGenresByBookUid(bookUid, utilityService.convertToGenreDomain(genres));
+    }
+
+    @Override
+    public void deleteGenresByBookUid(long bookUid, List<GenreDto> genres) {
+        genreDao.deleteGenresByBookUid(bookUid, utilityService.convertToGenreDomain(genres));
+    }
+
+    @Override
+    public List<GenreDto> getGenresByBookUid(long bookUid) {
+        return utilityService.convertToGenreDto(genreDao.getGenresByBookUid(bookUid));
     }
 }
