+++ homework-05/src/test/java/ru/otus/homework/service/BookGenreServiceImplTest.java
@@ -30,22 +30,20 @@ class BookGenreServiceImplTest {
     private final int EXPECTED_VALUE = 1;
 
     @Autowired
-    private BookGenreServiceImpl bookGenreService;
-    @Autowired
     private BookService bookService;
     @Autowired
     private GenreService genreService;
 
     @Test
     void shouldInsertGenresByBookUid() {
-        int size = bookGenreService.getGenresByBookUid(BOOK_UID).size();
+        int size = genreService.getGenresByBookUid(BOOK_UID).size();
         GenreDto genre0 = genreService.insert(new GenreDto(GENRE_NAME_0));
         GenreDto genre1 = genreService.insert(new GenreDto(GENRE_NAME_1));
         List<GenreDto> genres = new ArrayList<>();
         genres.add(genre0);
         genres.add(genre1);
-        bookGenreService.insertGenresByBookUid(BOOK_UID, genres);
-        assertEquals(size + 2, bookGenreService.getGenresByBookUid(BOOK_UID).size());
+        genreService.insertGenresByBookUid(BOOK_UID, genres);
+        assertEquals(size + 2, genreService.getGenresByBookUid(BOOK_UID).size());
     }
 
     @Test
@@ -55,26 +53,26 @@ class BookGenreServiceImplTest {
         List<GenreDto> genres = new ArrayList<>();
         genres.add(genre0);
         genres.add(genre1);
-        bookGenreService.editGenresByBookUid(BOOK_UID, genres);
-        assertEquals(EXPECTED_GENRES_COUNT_AFTER_EDIT, bookGenreService.getGenresByBookUid(BOOK_UID).size());
+        genreService.editGenresByBookUid(BOOK_UID, genres);
+        assertEquals(EXPECTED_GENRES_COUNT_AFTER_EDIT, genreService.getGenresByBookUid(BOOK_UID).size());
     }
 
     @Test
     void shouldDeleteGenresByBookUid() {
-        List<GenreDto> genres = bookGenreService.getGenresByBookUid(BOOK_UID);
+        List<GenreDto> genres = genreService.getGenresByBookUid(BOOK_UID);
         int size = genres.size();
         genres.remove(0);
-        bookGenreService.deleteGenresByBookUid(BOOK_UID, genres);
-        assertEquals(EXPECTED_VALUE, bookGenreService.getGenresByBookUid(BOOK_UID).size());
+        genreService.deleteGenresByBookUid(BOOK_UID, genres);
+        assertEquals(EXPECTED_VALUE, genreService.getGenresByBookUid(BOOK_UID).size());
     }
 
     @Test
     void shouldGetBooksByGenreUid() {
-        assertNotEquals(UNEXPECTED_VALUE, bookGenreService.getBooksByGenreUid(GENRE_UID));
+        assertNotEquals(UNEXPECTED_VALUE, bookService.getBooksByGenreUid(GENRE_UID));
     }
 
     @Test
     void shouldGetGenresByBookUid() {
-        assertNotEquals(UNEXPECTED_VALUE, bookGenreService.getGenresByBookUid(BOOK_UID));
+        assertNotEquals(UNEXPECTED_VALUE, genreService.getGenresByBookUid(BOOK_UID));
     }
 }
