+++ spring-shell.log
@@ -441,3 +441,47 @@
 1578072787692:59
 1578072790625:n
 1578072799565:show-books
+1578077225720:show-authors
+1578077236966:show-genres
+1578077249981:add-genre
+1578077257733:add-author
+1578077272703:Yet another brand new author
+1578077321430:add-author
+1578077326278:Yet another brand new author
+1578077362771:add-author
+1578077369623:Yet another brand new author
+1578077374556:His pen name
+1578077383438:show-authors
+1578077394629:show-books
+1578077407769:edit-book
+1578077415037:27
+1578077419739:20
+1578077421905:y
+1578077424353:21
+1578077429733:59
+1578077432352:y
+1578077434248:60
+1578077450007:Brand New Book (edited)
+1578077460264:12345678910
+1578077462471:2020
+1578077468294:show-books
+1578077482288:add-book
+1578077492111:26
+1578077499050:59
+1578077501690:60
+1578077504528:Crap
+1578077512621:10987654321
+1578077516002:2019
+1578077521535:show-books
+1578077536112:delete-author
+1578077538088:y
+1578077545295:59
+1578077552228:show-books
+1578077559151:edit-book
+1578077564805:28
+1578077572057:24
+1578077578129:60
+1578077582963:y
+1578077584714:56
+1578077587300:n
+1578077595888:show-books
